gmyc.support {splits}	R Documentation

Calculating AIC-based support values for the GMYC clusters

Description

This function calculates support values of species clusters delmited by the GMYC function using information-theoretic multimodel inference.

Usage


gmyc.support(object, p=0.95)
confset.gmyc(object, p=0.95)

Arguments

object
a gmyc object obtained by gmyc

p
level of confidence. models within p% confidence set are used to calculate support values

Details

The function gmyc.support calculates support values of the GMYC-delimited species by using the multimodel comparison approach described by Burnham & Anderson (2002). The support value of a node is defined as the sum of Akaike weights of candidate delimitation models in which the node is included. Only models included in the p% confidence set obtained by confset.gmyc are used for calculation. See reference papers for further details.

Value

gmyc.support returns a vector containing GMYC support values for tree nodes. Support values are ordered following the node order of input tree. confset.gmyc returns a gmyc object which contains models within p%confidence set.

Author(s)

Tomochika Fujisawa t.fujisawa05@gmail.com

References

Fujisawa, T. & Barraclough, TG. in press. Delimiting species using single-locus data and the generalized mixed Yule coalescent (GMYC) approach: A revised method and evaluation on simulated datasets. Systematic Biology.

Model selection and multimodel inference: A practical information-theoritic approach. 2nd ed. 2002. Springer Science. New York, USA.

Examples


data(test.tr)
test <- gmyc(test.tr, method="single")
nsup <- gmyc.support(test)

#plot support values on tree
plot(test)
nsup[nsup==0] <- NA
nodelabels(round(nsup, 2), cex=0.6, frame="n", adj=1)

#how many models within the confidence set?
cs <- confset.gmyc(test)
length(cs$MRCA)

[Package splits version 1.0-19 Index]