splits-defunct {splits}	R Documentation

Defunct functions in packaga splits

Description

Following function(s) has been removed from splits and no longer available.

Usage

compare(...)

Details

compare has been removed bacause the chi-square test between the single and multiple threshold models is not valid. (Their numbers of parameters are identical. See reference).

References

Fujisawa, T and Barraclough T.G. in press. Delimiting species using single locus data and the generalized mixed Yule coalescent approach. Systematic Biology.

[Package splits version 1.0-19 Index]